684 - finish presentational parts of password step
